MainWindow.cpp: Fix center position of flarm radar, even with asymmetric infobox-layout
The center positions of the flarm radar was centered on the screen, but did not reflect the position of the infoboxes. With eg. "8 Bottom or Right" in landscape, the flarm radar in the center of the screen covered a part of the inner infoboxes. The new formula fix that by using the edges of the map window (GetMainRect().x) instead the edges of the screen (rc.x).