MainWindow.cpp: MainWindow.cpp: Adapt size of flarm radar and thermal assistant to avoid overlap
Currently, the size of the flarm radar and the thermal assistant is set as the smaller value of the height or width of 2 infoboxes. With the setting "avoid infoboxes", and a map window smaller than 4 infoboxes, both gauges could overlap.

To improve this situation, the size of the gauges is now calculated as the smaller value of either the size of 2 infoboxes, or 50% of the size of the map window (smaller value of height or width also).

An overlap of the two gauges could only arrive if the flarm radar is set to CENTER, and the map window smaller than 4 infoboxes.

The formulae for flarm radar are now calculated with the same references (layout.control_size) as the thermal assistant.
'sz' is the min of height or width of 2 infoboxes, used already with the thermal assistant,
'mw' is the min of height or width of the map window,
'dia' is the min of 'sz' or 50% of 'mw', the calculated diameter of the gauges.