New workflow for autocommit translations
The workflow checks, if there are new strings to translate. If so, it provides an update of the files in po/, which then can be pulled from weblate. New translations are later pushed from weblate into the *.po files in po/.

To decide if the *.pot and *.po files shall be updated, a check is done on xcsoar.pot.
Before update of all files, a copy of xcsoar.pot is made (xcsoar.old), then the 'make update-po' is run.
Then, another copy of the updated xcsoar.pot is made (xcsoar.ref). In this file, and the old file, the line with the timestamp
is removed, before both files will be compared for differences.
If a difference is found, the local changes will be autocommited to the repository.